HACK: libpayload: Remove unnecessary keyboard mode setting code.
keyboard_init attempts to read the existing mode register, set the
'XLATE' bit, and write it back. The implementation is buggy because the
keyboard may be active at the time we read the mode, and we can
misinterpret scancode data as the reply to our command. It leads to
problems where the KB gets disabled in firmware.

(Source of change:
http://www.coreboot.org/pipermail/coreboot/2008-April/033899.html)

In fact, setting the 'XLATE' bit is completely unnecessary, even if we
desire QEMU keyboard support. We already set this bit when we initialize
the keyboard in pc_keyboard_init. Also, we have this bit set by default in
Chrome OS EC. Basically, this code does nothing (or worse), so just
remove it.

This change has not yet landed to master, nor to upstream Coreboot, so I
am labeling it 'HACK' (though I do not consider it a hack, in any
real definition of the word). Once it lands upstream, I will revert this
commit and take the upstreamed version.

Upstream progress can be tracked here:

http://review.coreboot.org/#/c/3883/

BUG=chrome-os-partner:22134
TEST=Manual on Peppy. Spam keyboard going into recovery mode, verify the
keyboard still remains functional. Verify keyboard functions in dev
mode, recovery mode, and verified boot.
BRANCH=FalcoPeppy.

Change-Id: Id9ec61d67790131ec829b2b9f80c727d6f6acd51
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66918
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>