peppy: Re-read critical temperatures in ACPI _TMP
There seem to be a significant number of shutdowns during suspend resume
tests related to critical temperatures.  It is possible that we are getting
a bad reading from PECI and shutting down prematurely in some cases.

If we get a reading that is above critical then wait for the EC to re-poll
and then re-check the temperature in case it was just a bad reading.

Also add some ACPI debug messages when this happens.

BUG=chrome-os-partner:19980
BRANCH=peppy
TEST=Successful 10000 suspend_stress_test on 4 falco devices

Change-Id: I46c281fba50ef0ddc4e7f4c7e83378adea49ed93
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/167652
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>