Leon: Delay the setup of pc-beep verbs
Leon has an unwanted  audio pop sound on system warm reboots.
The cause is that the ALC283 requires about 100ms between the
device reset and the setup of the pc-beep verbs. Specificly the
EAPD signal enable needs to be delayed. We don't want to wait
100ms in the azalia code, so add a callback to setup the pc-beep
verb after the coreboot table setup.

BUG=chrome-os-partner:24036
BRANCH=leon
TEST=Manual. From vt2 use the reboot command and listen for
a pop sound prior to the dev screen being visible.

Change-Id: If75dc6aa12a1a5599604024ca84fb3973c92f8f7
Signed-off-by: Marc Jones <marc.jones@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/180984
Reviewed-by: Aaron Durbin <adurbin@chromium.org>