squawks : remove I2C6 and the associated atmel touch devices
Squawks will not have a touchscreen, and this bus is actually left
floating, so lets turn off the bus and remove the entries for the
atmel touchscreen that exists on some Rambis but not Squawks.

Note, I2C6 is the name of this bus in coreboot where the busses
are 1 indexed, but on the 0 indexed schematic, this is i2c5.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:26010
TEST=Build and deploy firmware for squawks.
check that i2cdetect -y -l only lists three busses named
"Synopsys DesignWare I2C adapter"
Also check with i2cdetect -y -r with busses 0, 1, 2 that
these busses don't randomly show devices on these busses.

Change-Id: I7a694742c828def52bd95026aa0a256b0919fd32
Reviewed-on: https://chromium-review.googlesource.com/187170
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>