monroe: Change critical shutdown temperature to 98C
Due to occasional SIO temperature reading flakes, it's not currently
possible to distinguish 100C temp readings from temperature read errors.
Lower the critical shutdown temperature to 98C so that it's possible for
the system to actually shutdown. We should revert this change once we
have fixed the root cause of the temperature read errors.

BUG=None.
TEST=Manual on Monroe. Stress test device with no fan attached, verify
that system gracefully goes to thermal shutdown.

Change-Id: I950d895e2187e0d2f6bf9852e2435a4661917aea
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/190695