publish lumpy/stumpy firmware
This is a snapshot from:
	https://chrome-internal.googlesource.com/chromeos/third_party/coreboot
at git sha1:
	2a5362c5bf2e070f1b312317666404fac2c12117

All files in the 3rdparty/ subdir have been stripped.

Further, swap the internal base from:
	1f0b0d9ffb318b7869abb06b9213f5c231ac0e57
to the public base:
	8eee19d0eafd3a34742df4d26c810424097211fe

They contain the same source, but the svn-to-git trees diverged at commit:
	7272fcb2a466d2554b29e082c8e57038bfc4ebe6
The differences are limited to metadata in git itself and not the files
the commits contain.

BUG=None
TEST=compiled internal repo & public repo and compared disassembled files

Change-Id: Ic63d6647474c7a527d41886b03a66ce1316e1a65
Reviewed-on: https://chromium-review.googlesource.com/194379
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>