nyan*: enable CLAMP_INPUTS
Enable pinmux clamp function to avoid pinmux conflict.
For pins which are configured to tristate enabled, the inputs to the
controller will be clamped to zero. This can be used to avoid pinmux
conflicts since the tristate bit is set to 1 in the power-on-reset
pinmux setting.
With pinmux clamp enabled, we need to configure all the input pins
to tristate disabled.
BUG=chrome-os-partner:27091
BRANCH=None
TEST=built and booted successfully, display worked fine.

Signed-off-by: Ken Chang <kenc@nvidia.com>
(cherry picked from commit c95d6fe79810612cfad721667657cdcb87068d23)

Change-Id: Ia8e27053e4ed4f10e69f57b0c67efa8823852867
Reviewed-on: https://chromium-review.googlesource.com/196423
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>