CHROMIUM: coreboot: Program i2c-designware SDA Hold Register
I2C bus SDA hold time can be marginal with 60ns value, especially
when there is level shifter on the bus. So program it to 300ns
based on Fast-mode specification, which is between 0 to 900ns.
Apply the same timing for Standard-mode as well.

BUG=chrome-os-partner:28886
TEST=Manual: Check i2c-designware drivers work on Rambi.

Change-Id: I799ec066eba5ebb8134a5fe7a479c25a603cef1b
Reported-by: dusonlin@emc.com.tw
Signed-off-by: Wenkai Du <wenkai.du@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/201448
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>
Commit-Queue: Keith Tzeng <keith.tzeng@quantatw.com>
Tested-by: Keith Tzeng <keith.tzeng@quantatw.com>