zako: Initialize display when we need to show the "incorrect adapter" screen.
Before displaying the "incorrect adapter", we have to initialize Coreboot
display and frame buffer, which were both skipped when system is running in
normal mode.

To solve that (and to make sure we don't need a RO update), added
"get_need_display" in per-board implementation files and new "cros_need_display"
for VBIOS and display subsystems to know if they should be executed.

For Zako, the get_need_display is done by reading and checking adapter type.
Note the real screen must be displayed by payload.

BRANCH=zako
BUG=chrome-os-partner:29391
TEST=emerge-zako chromeos-coreboot-zako chromeos-bootimage
     Set DUT to normal mode, plug 45W adapter, and reboot.
     See warning screen correctly.
     Power off, then power on again, still seeing warning screen.
     Power off, change to 65W adapter, and reboot successfully.

Change-Id: I7f2d1b469e9d76c9e4a86031841b9f21e860a374
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203364
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>